added support for array_set in the full-slice option

+++ regression/goto-instrument/chain.sh
@@ -12,15 +12,20 @@ name=${name%.c}
 
 args=${@:1:$#-1}
 
-$goto_cc -o $name.gb $name.c
+if [ -f $name.c ]
+then
+  $goto_cc -o $name.gb $name.c
 # $goto_instrument --show-goto-functions $name.gb
-$goto_instrument $args $name.gb ${name}-mod.gb
-if [ ! -e ${name}-mod.gb ] ; then
-  cp $name.gb ${name}-mod.gb
-elif echo "$args" | grep -q -- "--dump-c" ; then
-  mv ${name}-mod.gb ${name}-mod.c
-  $goto_cc ${name}-mod.c -o ${name}-mod.gb
-  rm ${name}-mod.c
+  $goto_instrument $args $name.gb ${name}-mod.gb
+  if [ ! -e ${name}-mod.gb ] ; then
+    cp $name.gb ${name}-mod.gb
+  elif echo "$args" | grep -q -- "--dump-c" ; then
+    mv ${name}-mod.gb ${name}-mod.c
+    $goto_cc ${name}-mod.c -o ${name}-mod.gb
+    rm ${name}-mod.c
+  fi
+else
+  $goto_instrument $args $name ${name}-mod.gb
 fi
 $goto_instrument --show-goto-functions ${name}-mod.gb
 $cbmc ${name}-mod.gb
